backend/lambda/python/requirements.txt
scipy
numpy
SpeechRecognition
ffmpeg-python